Harden codebase with bounds checks, NULL guards, and error path fixes
## Summary
- Add bounds checks to packfile parser: read_type_and_size, apply_delta COPY/INSERT, REF_DELTA SHA read, and memchr NULL check
- Fix clone_repo to restore chdir on all error paths via goto cleanup pattern
- Fix checkout_tree to return errors on truncated tree entries instead of silent break
- Add NULL input guards to decompress_data, hex_to_string, hex_string_to_bytes, object_path
- Fix sprintf→snprintf in hex_to_string, overflow guard in decompress_data initial size
- Replace directory_exists+mkdir TOCTOU with atomic mkdir+EEXIST check in object_write
- Rename header guards STRING_H→GIT_STRING_H, FILE_H→GIT_FILE_H to avoid system collisions
- Change GitObject.body_size from unsigned long to size_t for type consistency
- Add CONFIGURE_DEPENDS to CMake GLOB_RECURSE and libcurl to WIN32 build

## Test plan
- [ ] Verify all 7 CodeCrafters stages pass (init, cat-file, hash-object, ls-tree, write-tree, commit-tree, clone)
- [ ] Verify clone works against codecrafters-io/git-sample repos

+++ src/objects/object.h
@@ -14,7 +14,7 @@
 /* Parsed git object — body points into raw, so only raw needs freeing. */
 typedef struct {
     unsigned char *body;       /* object content (past the header + \0) */
-    unsigned long  body_size;  /* byte count of body */
+    size_t         body_size;  /* byte count of body */
     unsigned char *raw;        /* full decompressed buffer (caller frees) */
 } GitObject;
